Refactor logging in csidriver package
